requirements.txt
torch>=2.1
pyyaml>=6.0
numpy>=1.23
